+++ crates/swc_ecma_transforms_module/src/import_rewriter_typescript.rs
