+++ crates/swc_ecma_transforms_module/src/lib.rs
@@ -14,13 +14,13 @@ pub mod util;
 pub mod amd;
 pub mod common_js;
 pub mod import_analysis;
+pub mod import_rewriter_swc;
+pub mod import_rewriter_typescript;
 pub(crate) mod module_decl_strip;
 pub(crate) mod module_ref_rewriter;
 pub mod path;
-pub mod rewriter;
 pub mod system_js;
 mod top_level_this;
-pub mod typescript_import_rewriter;
 pub mod umd;
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize)]
